# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CallAssistanceRequest:
    properties:
      description:
        type: string
      location:
        type: string
      rental_id:
        type: integer
    required:
    - description
    - location
    - rental_id
    type: object
...
  handlers.PaymentRequest:
    properties:
      rental_id:
        type: integer
    type: object
...
  handlers.TopUpRequest:
    properties:
      deposit_amount:
        type: number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is Jakarta Luxury Rent Car service API documentation.
  title: Jakarta Luxury Rent Car API
  version: "1.0"
paths:
...
  /users/call-assistance:
    post:
      consumes:
      - application/json
      description: Call assistance request
      parameters:
      - description: Call assistance request body containing rental ID, location,
          and description
        in: body
        name: assistanceReq
        required: true
        schema:
          $ref: '#/definitions/handlers.CallAssistanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success message and details of the call assistance request
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format or validation error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User, rental history, or car not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create call assistance record or send WhatsApp notification
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Call assistance request
      tags:
      - Role User
  /users/get-deposit:
    get:
      consumes:
      - application/json
      description: RGet user deposit amount
      produces:
      - application/json
      responses:
        "200":
          description: User's email and deposit amount
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user deposit amount
      tags:
      - Role User
  /users/get-membership:
    get:
      consumes:
      - application/json
      description: Get user membership details
      produces:
      - application/json
      responses:
        "200":
          description: Membership details including membership ID, user ID, email,
            and discount level
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found or no membership found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user membership details
      tags:
      - Role User
  /users/making-payment:
    post:
      consumes:
      - application/json
      description: MakingPayment updates status from "Book" to "Paid"
      parameters:
      - description: Payment request body containing rental ID and payment details
        in: body
        name: paymentReq
        required: true
        schema:
          $ref: '#/definitions/handlers.PaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rental status updated to 'Paid' successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or rental status is not 'Book'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Rental history or user not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update rental status or deposit amount
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: MakingPayment updates status from "Book" to "Paid"
      tags:
      - Role User
  /users/register-membership:
    post:
      consumes:
      - application/json
      description: Register a new membership
      produces:
      - application/json
      responses:
        "200":
          description: Details of the newly registered membership including membership
            ID, user ID, email, and discount level
          schema:
            additionalProperties: true
            type: object
        "400":
          description: User already has a membership registered
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to register membership
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Register a new membership
      tags:
      - Role User
  /users/topup:
    post:
      consumes:
      - application/json
      description: Top up user deposit amount
      parameters:
      - description: Top-up request body
        in: body
        name: topUpReq
        required: true
        schema:
          $ref: '#/definitions/handlers.TopUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated email and deposit amount
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update deposit amount
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Top up user deposit amount
      tags:
      - Role User
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"